backend/requirements.txt
Pillow
requests
django-debug-toolbar
django-jazzmin
channels 
channels_redis 
Django>=5.0
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
django-cors-headers
google-genai>=1.0.0
gunicorn
